reorg to get rid of src directory
