+++ rampr/datasets/__init__.py
@@ -1,0 +1,21 @@
+from .core import (
+    DatasetError,
+    fetch,
+    get_spec,
+    get_path_by_filename,
+    info,
+    list_datasets,
+    path,
+    list_dataset_files,
+)
+
+__all__ = [
+    "DatasetError",
+    "fetch",
+    "get_spec",
+    "get_path_by_filename",
+    "info",
+    "list_datasets",
+    "list_dataset_files",
+    "path",
+]
