--- src/rampr/datasets/__init__.py
@@ -1,10 +1,0 @@
-from .core import DatasetError, fetch, get_spec, info, list_datasets, path
-
-__all__ = [
-    "DatasetError",
-    "fetch",
-    "get_spec",
-    "info",
-    "list_datasets",
-    "path",
-]
